requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.44.0
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.1
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.4.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
